+++ framework/core/src/libkungfu/wingchun/operator/backtest.cpp
@@ -247,7 +247,7 @@ void BacktestContext::subscribe_slice(const location_ptr &slice_location, int64_
       return;
     }
   };
-  int lead_ratio = from_indexer_->acquire_lead_ratio();
+  float lead_ratio = from_indexer_->acquire_lead_ratio();
   if (lead_ratio < 0) {
     throw wingchun_error(fmt::format("lead_ratio={} is invalid", lead_ratio));
   }
@@ -255,7 +255,7 @@ void BacktestContext::subscribe_slice(const location_ptr &slice_location, int64_
 }
 
 void BacktestContext::unsubscribe_slice(const location_ptr &slice_location, int64_t nanotime, int64_t offset) {
-  int delay_ratio = from_indexer_->release_delay_ratio();
+  float delay_ratio = from_indexer_->release_delay_ratio();
   if (delay_ratio < 0) {
     throw wingchun_error(fmt::format("delay_ratio={} is invalid", delay_ratio));
   }
